:construction_worker: alimentation Bdd
