+++ src/main/java/com/scc/controller/AdminController.java
@@ -1,0 +1,45 @@
+package com.scc.controller;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+import javax.transaction.Transactional;
+import javax.validation.Valid;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+
+import com.scc.model.PgArbreGenealogie;
+import com.scc.model.PgDog;
+import com.scc.service.DogService;
+
+@Path("/public/admin")
+@ApplicationScoped
+@Consumes(MediaType.APPLICATION_JSON)
+@Produces(MediaType.APPLICATION_JSON)
+public class AdminController {
+
+    @Inject
+    DogService dogService;
+    
+    @POST
+    @Transactional
+    @Path("/pedigrees")
+    public Response create(@Valid PgDog dog){
+        
+        dogService.populateDog(dog);
+        return Response.status(Response.Status.CREATED).entity(dog).build();
+    }
+
+    @POST
+    @Transactional
+    @Path("/parents")
+    public Response create(@Valid PgArbreGenealogie parents){
+        
+        dogService.populateParents(parents);
+        return Response.status(Response.Status.CREATED).entity(parents).build();
+    }
+}
