+++ src/main/java/com/scc/service/LoginAttemptService.java
@@ -1,0 +1,74 @@
+package com.scc.service;
+
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+
+import javax.annotation.PostConstruct;
+import javax.enterprise.context.ApplicationScoped;
+
+import org.eclipse.microprofile.config.inject.ConfigProperty;
+import org.jboss.logging.Logger;
+
+import com.google.common.cache.CacheBuilder;
+import com.google.common.cache.CacheLoader;
+import com.google.common.cache.LoadingCache;
+
+@ApplicationScoped
+public class LoginAttemptService {
+
+    private static final Logger LOG = Logger.getLogger(LoginAttemptService.class);
+    
+    @ConfigProperty(name = "access.max.attempt")    
+    int MAX_ATTEMPT;
+    @ConfigProperty(name = "access.recovery.attempt")    
+    int DELAY_ATTEMPT;
+    
+    private boolean ACTIVE_LOCK;
+    private LoadingCache<String, Integer> attemptsCache;
+    
+    @PostConstruct
+    public void init() {
+        ACTIVE_LOCK = (MAX_ATTEMPT == -1 ? false : true);
+        
+        attemptsCache = CacheBuilder.newBuilder().expireAfterWrite(DELAY_ATTEMPT, TimeUnit.MINUTES).build(new CacheLoader<String, Integer>() {
+            @Override
+            public Integer load(final String key) {
+                return 0;
+            }
+        });
+    }
+
+    public void loginSucceeded(final String key) {
+        if (!ACTIVE_LOCK)
+            return;
+        
+        attemptsCache.invalidate(key);
+    }
+    
+    public void loginFailed(final String key) {
+
+        if (!ACTIVE_LOCK)
+            return;
+        
+        int attempts = 0;
+        try {
+            attempts = attemptsCache.get(key);
+        } catch (final ExecutionException e) {
+            attempts = 0;
+        }
+        attempts++;
+        attemptsCache.put(key, attempts);
+    }
+
+    public boolean isBlocked(final String key) {
+        
+        if (!ACTIVE_LOCK)
+            return false;
+
+        try {
+            return attemptsCache.get(key) >= MAX_ATTEMPT;
+        } catch (final ExecutionException e) {
+            return false;
+        }
+    }    
+}
